moved construction paste into separate package, misc cleanup
